+++ internal/utils/registry/teams/teams.go
@@ -1,11 +1,12 @@
 package teams
 
 const (
-	Protocol   = `protocol`
-	UI         = `ui`
-	CSM        = `csm`
-	Governance = `governance`
-	ValSet     = `valset`
-	Analytics  = `Analytics`
-	DevOps     = `DevOps`
+	Protocol      = `protocol`
+	ProtocolStage = `protocolStage`
+	UI            = `ui`
+	CSM           = `csm`
+	Governance    = `governance`
+	ValSet        = `valset`
+	Analytics     = `Analytics`
+	DevOps        = `DevOps`
 )