+++ Changelog.md
@@ -1,3 +1,6 @@
+## 09.10.2024
+1. Added stage consumers
+
 ## 08.10.2024
 1. Split up feeder, forwarder to independent bin applications
 2. Added mechanism for using UniqueKey for collecting quorum